\name{Background}
\alias{Background}
\title{
  Background
}
\description{
  Returns the background noise associated with the panel
}
\usage{
Background(geneNames,
           samplesNormalizedReadCounts,
           referenceNormalizedReadCounts,
           sigList,
           replicates = 1000)
}
\arguments{
  \item{geneNames}{
    Names of the genes
}
  \item{samplesNormalizedReadCounts}{
  A matrix with the normalized read counts
}
  \item{referenceNormalizedReadCounts}{
  A matrix with the reference read counts
}
  \item{sigList}{
  List of significant genes
}
  \item{replicates}{
    List of replicates
}
}
\details{
TODO create some
}
\value{
TODO create some
}
\references{
TODO create some
}
\author{
  Cristiano Oliveira, Thomas Wolf
}
\note{
TODO create some
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
 %%objects to See Also as \code{\link{help}},
 TODO create some
}
\examples{
%%TODO create some
2 + 2

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
