\name{BackgroundReport}
\alias{BackgroundReport}
\title{
  BackgroundReport
}
\description{
Generate a report from the background noise
}
\usage{
BackgroundReport(background, geneNames)
}

\arguments{
  \item{background}{
Background noise associated with the genes
}
  \item{geneNames}{
Names of the genes
}
}
\details{
TODO create some
}
\value{
Returns a report based on the background noise
}
\references{
TODO create some
}
\author{
Cristiano Oliveira, Thomas Wolf
}
\note{
TODO create some
}
\seealso{
%% objects to See Also as \code{\link{help}},
TODO create some
}
\examples{
%% TODO create some
2 + 2
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
