\name{BootList}
\alias{BootList}
\title{
  BootList
}
\description{
  TODO Ask Thomas!!!
  A concise (1-5 lines) description of what the function does.
}
\usage{
BootList(geneNames, sampleMatrix, refmat, reps, refWeights)
}
\arguments{
  \item{geneNames}{
     Describe \code{geneNames} here
}
  \item{sampleMatrix}{
     Describe \code{sampleMatrix} here
}
  \item{refmat}{
     Describe \code{refmat} here
}
  \item{reps}{
     Describe \code{reps} here
}
  \item{refWeights}{
     Describe \code{refWeights} here
}
}
\details{
  If necessary, more details than the description above
}
\value{
# TODO ask Thomas
  Describe the value returned
  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
 ...
}
\references{
TODO create some
}
\author{
  Thomas Wolf, Cristiano Oliveira
}
\note{
  TODO create some
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
TODO create some
 objects to See Also as \code{\link{help}},
}
\examples{
%% TODO create some
# bootList = BootList(genesPositionsIndex,
#                    samplesNormalizedReadCounts,
#                    referenceNormalizedReadCounts,
#                    reps = replicates,
#                    refWeights = NULL)


}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
