\name{CheckSignificance}
\alias{CheckSignificance}
\title{
  CheckSignificance
}
\description{
    This function uses the bootstrap/subsampling distribution of the ratio between the mean
    averaged reference read counts and the samples of interest. Based on the 95 percent
    confidence interval of this distribution, the significance of a CNV aberration being present
    is decided.
}
\usage{
CheckSignificance(bootList)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{bootList}{
      A list as generated by the BootList function
  }
}
\value{
A list of report matrices. One entry for each entry of the bootList parameter.
Each matrix contains the lower bound and the upper bound of the 95 percent confidence
interval and logical information about the significance status.
}

\author{
    Thomas Wolf, Cristiano Oliveira
}

\examples{
    \donttest{
        CheckSignificance(bootList)
    }
}

