\name{geneNames}
\docType{data}
\alias{geneNames}
\title{Gene Names}
\description{
  This data set provides gene names to be used as part of CNVPanelizer sample data.
}
\usage{geneNames}
\format{A vector}
\source{Automatically generated data}
\references{
TODO
  Wolf, D. R. (2015) \emph{CNVPanelizer}.
  New York: Wiley.
}
\keyword{datasets}
