\name{genomicRanges}
\docType{data}
\alias{genomicRanges}
\title{Genomic Ranges}
\description{
  A genomic ranges object to be used as part of CNVPanelizer sample data.
}
\usage{genomicRanges}
\format{A genomicRanges object}
\source{Automatically generated data}
\references{
TODO
  McNeil, D. R. (1977) \emph{Interactive Data Analysis}.
  New York: Wiley.
}
\keyword{datasets}