\name{PlotBootstrapDistributions}
\alias{PlotBootstrapDistributions}
\title{
  PlotBootstrapDistributions
}
\description{
  Plots the generated bootstrap distribution
}
\usage{
PlotBootstrapDistributions(bootList, reportTables, outputFolder)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{bootList}{
      List of bootstrapped read counts for each sample data
  }
  \item{reportTables}{
      List of report tables for each sample data
  }
  \item{outputFolder}{
    Path to the folder where the data plots will be created
  }
}
\details{
TODO create some
  If necessary, more details than the description above
}
\value{
TODO create some
  Describe the value returned
  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
 ...
}
\references{
TODO create some
}
\author{
  Thomas Wolf, Cristiano Oliveira
}
\note{
  TODO create some
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
TODO create some
 objects to See Also as \code{\link{help}},
}
\examples{
# PlotBootstrapDistributions(bootList, reportTables, outputFolder)

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
