\name{referenceReadCounts}
\docType{data}
\alias{referenceReadCounts}
\title{Reference read data sample}
\description{
  This data set gives read counts from some sequenced sample.
}
\usage{referenceReadCounts}
\format{A matrix}
\source{Automatically generated data}
\references{
TODO
  McNeil, D. R. (1977) \emph{Interactive Data Analysis}.
  New York: Wiley.
}
\keyword{datasets}