\name{sampleReadCounts}
\docType{data}
\alias{sampleReadCounts}
\title{Test sample data}
\description{
    Synthetic data set for simulating read counts
    from a Panel sequencing experiment
}
\usage{testSamples}
\value{A matrix with columns identifying the sample names and
    columns the gene names}
\format{A matrix with columns identifying the sample names and
    columns the gene names}
\source{Artificially generated data}
\keyword{datasets}

