\name{SignificantGenes}
\alias{SignificantGenes}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  SignificantGenes
}
\description{
  TODO create some
  A concise (1-5 lines) description of what the function does.
}
\usage{
SignificantGenes(sigList, genesPositionsIndex)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sigList}{
      TODO create some
     Describe \code{sigList} here
  }
  \item{genesPositionsIndex}{
      TODO create some
     Describe \code{genesPositionsIndex} here
  }
}
\details{
TODO create some
  If necessary, more details than the description above
}
\value{
TODO create some
  Describe the value returned
  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
 ...
}
\references{
TODO create some
}
\author{
  Thomas Wolf, Cristiano Oliveira
}
\note{
  TODO create some
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
TODO create some
 objects to See Also as \code{\link{help}},
}
\examples{
%% TODO create some
2 + 2

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
