\name{CheckSignificance}
\alias{CheckSignificance}
\title{
  CheckSignificance
}
\description{
  TODO Ask Thomas   A concise (1-5 lines) description of what the function does.
  Calculate the significance from the bootstrap sampling
}
\usage{
CheckSignificance(bootList)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{bootList}{
      TODO create some
     Describe \code{bootList} here
  }
}
\details{
TODO create some
  If necessary, more details than the description above
}
\value{
TODO create some
  Describe the value returned
  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
 ...
}
\references{
TODO create some
}
\author{
  Thomas Wolf, Cristiano Oliveira
}
\note{
  TODO create some
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
TODO why is this not highlighted?
  \code{\link{BootList}}
}
\examples{
%% TODO create some
#sigList = CheckSignificance(bootList)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
