\name{CheckSignificance}
\alias{CheckSignificance}
\title{
  CheckSignificance
}
\description{
    This function compares the ratio between the median reference and
    the samples of interest, against the bootstrap subsampling distribution
    generated for each gene. Based on the 95 percent confidence interval the
    significance of a CNV aberration being present is being decided.
}
\usage{
CheckSignificance(bootList)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{bootList}{

      A list as generated by the BootList function
  }
}
\value{

A list of report matrices. One entry for each entry of the bootList parameter.
Each matrix contains the 5 percent quantile, the 95 percent quantile and logical
information about the significance status.
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
 ...
}

\author{
    Thomas Wolf, Cristiano Oliveira
}



\examples{

\dontrun{
    CheckSignificance(bootList)
}
}

