\name{ReadXLSXToList}
\alias{ReadXLSXToList}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  ReadXLSXToList
}
\description{
  Reads a list of read count matrices from a xlsx as generated by WriteReadCountsToXLSX
}
\usage{
ReadXLSXToList(filepath)
}
%- maybe also 'usage' for other objects documented here.
%%TODO create some
\arguments{
  \item{filepath}{
	filepath
}
}

\value{

  A list of read count matrices
}

\author{
  Thomas Wolf, Cristiano Oliveira
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{

\dontrun{
 ReadXLSXToList(filepath)
}
}
