\name{ReadXLSXToList}
\alias{ReadXLSXToList}
\title{
  ReadXLSXToList
}
\description{
    Reads a list of read count matrices from a xlsx as generated
    by WriteReadCountsToXLSX
}
\usage{
ReadXLSXToList(filepath)
}

\arguments{
  \item{filepath}{
    filepath
  }
}

\value{
    A list of read count matrices
}

\author{
    Thomas Wolf, Cristiano Oliveira
}

\examples{
    \dontrun{
        ReadXLSXToList(filepath)
    }
}
