\name{referenceReadCounts}
\docType{data}
\alias{referenceReadCounts}
\title{Reference sample data}
\description{
	Synthetic reference data set for simulating read counts from a Panel sequencing experiment
}
\usage{referenceSamples}
\format{A matrix with columns identifying the sample names and columns the gene names}
\source{Artificially generated data}
\references{

}
\keyword{datasets}