\name{referenceReadCounts}
\docType{data}
\alias{referenceReadCounts}
\title{Reference sample data}
\description{
    Synthetic reference data set of simulated read counts.
    Only to be used for code examples.
}
\usage{referenceSamples}
\value{A matrix with columns identifying the sample names and
columns the gene names}
\format{A matrix with columns identifying the sample names and
columns the gene names}
\source{Artificially generated data}
\keyword{
    datasets
}
