\name{ReportTables}
\alias{ReportTables}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  ReportTables
}
\description{
  ReportTables
}
\usage{
ReportTables(bootList,
             geneNames,
             backgroundNoise,
             referenceNormalizedReadCounts,
             samplesNormalizedReadCounts,
             tableNames)
}
%% TODO create some argument list
\arguments{
  \item{bootList}{
     Describe \code{bootList} here
}
  \item{geneNames}{
     Describe \code{geneNames} here
}
  \item{backgroundNoise}{
     Describe \code{backgroundNoise} here
}
  \item{referenceNormalizedReadCounts}{
     Describe \code{referenceNormalizedReadCounts} here
}
  \item{samplesNormalizedReadCounts}{
     Describe \code{samplesNormalizedReadCounts} here
}
  \item{tableNames}{
     Describe \code{tableNames} here
}
}
\details{
TODO create some
}
\value{
  TODO create some
  Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}
 ...
}
\references{
TODO create some
}
\author{
TODO create some
}
\note{
TODO create some
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
  TODO create some
 objects to See Also as \code{\link{help}},
}
\examples{
%% TODO create some
2 + 2
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
