\name{sampleReadCounts}
\docType{data}
\alias{sampleReadCounts}
\title{Sample read data}
\description{
  This data set gives read counts from some sequenced sample.
}
\usage{sampleReadCounts}
\format{A matrix}
\source{Automatically generated data}
\references{
TODO
  McNeil, D. R. (1977) \emph{Interactive Data Analysis}.
  New York: Wiley.
}
\keyword{datasets}