\name{sampleReadCounts}
\docType{data}
\alias{sampleReadCounts}
\title{Test sample data}
\description{
    Synthetic data set of simulated read counts.
    Only to be used for running the code examples.
}
\usage{testSamples}
\value{A matrix with columns identifying the sample names and
    columns the gene names}
\format{A matrix with columns identifying the sample names and
    columns the gene names}
\source{Artificially generated data}
\keyword{datasets}

