\name{WriteReadCountsToXLSX}
\alias{WriteReadCountsToXLSX}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  WriteReadCountsToXLSX
}
\description{
  Writes both the read counts from reference and test samples to an xls file
}
\usage{
WriteReadCountsToXLSX(sampleReadCounts, referenceReadCounts, filepath = "readCounts.xls")
}
%- maybe also 'usage' for other objects documented here.
%%TODO create some
\arguments{
  \item{sampleReadCounts}{
    sampleReadCounts
}
  \item{referenceReadCounts}{
    referenceReadCounts
}
  \item{filepath}{
	filepath
}
}
\details{
TODO create some
}
\value{
TODO create some
  Describe the value returned
  If it is a LIST, use
  \item{sampleReadCounts}{Description of 'sampleReadCounts'}
  \item{referenceReadCounts}{Description of 'referenceReadCounts'}
  \item{{filepath}}{Description of 'filepath'}
 ...
}
\references{
TODO create some
}
\author{
  Thomas Wolf, Cristiano Oliveira
}
\note{
  TODO create some
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
 TODO create some
 objects to See Also as \code{\link{help}},
}
\examples{
%% TODO create some
2 + 2

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
